atc: behaviour: Use atc.VersionConfig in resource config
This switches the type of the `version:` field on a resource definition
to match the type of the `version:` field on a job `get:` definition.
This allows for the usage of 'latest' and 'every' as version on the
actual resource in line with #5456. This commit changes the type and
relevant usages and adds a test that `latest` can be correctly
deserialized. As the type in api.presenter has not changed, this does
not affect the API (and thus the UI at all). As the only part of the
resource versioning system that is used is `ConfigPinnedVersion` which
only references the `Pinned` field of the `atc.VersionConfig` flag, the
'latest' and 'every' variants have _no effect_ on the resource
versioning meaning this is effectively a no-op other than to allow
setting of `version: latest/every` in pipeline config

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>