atc: behaviour: Allow 'version: every' on resources
Add the necessary logic to the scheduler to check for `version: every`
that has been configured on a resource rather than just on a `get`. Add
tests to this effect (mirroring testflight/resource_version_test.go). It
would be preferable to be able to test this logic a bit lower down the
testing triangle but there are existing examples for these tests on the
`get` steps in testflight to match those

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>