atc/worker: Add new "total memory" placement strategy
Using the configured container limits for a task (or the global
defaults), use the amount of memory currently running on a worker to
decide if a task can can't be placed. This is similar to the active
tasks/active volumes strategies but attempts to be a bit more refined by
considering the size of tasks rather than just the number of them. Usage
of the ContainerRepository may mean for large installations (100s of
workers) that the strategy could be slow - if this appears to be the
case we could implement this instead as a property of the worker and
fetch it in a JOIN when we fetch all the compatible workers. Allow for
per-worker allocations so that the strategy can be used even when
workers are not a homogenous size

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>