atc/exec: Make "No workers" a retriable error
In distributed systems, workers may come and go - especially in
environments like Kubernetes where Pods may be recreated regularly due
to autoscaling etc. It would be nice if jobs did not fail due to not
having a worker and instead just waited until a worker was available.
This may lead to "thundering herd" issues but the "max containers" or
other similar strategies can help alleviate this

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>