atc + worker: Change limit-total-resources strategy to only memory
CPU entitlement is measured in shares which is a no-absolute value
(ie. there are no fixed amount of shares available, only the %age of the
total allocated shares controls how much CPU is available). As such, it
doesn't make much sense to try and limit the total (absolute) amount of
shares available on a worker. If CPU competition is an issue, another
strategy like limit-active-containers or limit-active-tasks can be
placed in the strategy chain to help with that

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>